Refactor test home

+++ indexlake/src/catalog/record.rs
@@ -13,8 +13,8 @@ pub(crate) struct DataFileRecord {
     pub(crate) data_file_id: i64,
     pub(crate) table_id: i64,
     pub(crate) relative_path: String,
-    pub(crate) file_size_bytes: usize,
-    pub(crate) record_count: usize,
+    pub(crate) file_size_bytes: i64,
+    pub(crate) record_count: i64,
 }
 
 impl DataFileRecord {
